HgScm: refactor HgRepo
[This commit is not functional on its own]
